requirements.txt
torch
torchvision
facenet-pytorch
onnxruntime
lpips
opencv-python
Pillow
numpy
scipy
tqdm
scikit-image
